refactor(server): encapsulate experimental extensions
