+++ apps/server/src/server.ts
@@ -54,15 +54,13 @@ import { createReadStream } from "node:fs";
 import { Readable } from "node:stream";
 import { serve, type ServeResult } from "./serve-node.js";
 import {
-  callGoogleWorkspaceExtensionAction,
   createGoogleWorkspaceConnectFlowManager,
-  GOOGLE_WORKSPACE_EXTENSION_ACTIONS,
-  GOOGLE_WORKSPACE_EXTENSION_ID,
   googleWorkspaceDisconnect,
   googleWorkspaceRunScopeSmokeTest,
   googleWorkspaceStatus,
   googleWorkspaceTestConnection,
-} from "./google-workspace.js";
+} from "./extensions/google-workspace.js";
+import { callExperimentalExtensionAction, listExperimentalExtensionActions } from "./extensions/index.js";
 import pkg from "../package.json" with { type: "json" };
 import constants from "../../../constants.json" with { type: "json" };
 
@@ -108,43 +106,8 @@ const OPENWORK_VOICE_REALTIME_TOOLS = [
   },
 ];
 
-const OPENWORK_EXPERIMENTAL_EXTENSION_ACTIONS = [
-  ...GOOGLE_WORKSPACE_EXTENSION_ACTIONS,
-];
-
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function listExperimentalExtensionActions(extensionId: string) {
-  const filter = extensionId.trim();
-  return filter
-    ? OPENWORK_EXPERIMENTAL_EXTENSION_ACTIONS.filter((action) => action.extensionId === filter)
-    : OPENWORK_EXPERIMENTAL_EXTENSION_ACTIONS;
-}
-
-async function callExperimentalExtensionAction(config: ServerConfig, input: unknown) {
-  if (!isRecord(input)) {
-    throw new ApiError(400, "invalid_payload", "Expected extension action call payload");
-  }
-  const extensionId = readStringField(input, "extensionId");
-  const action = readStringField(input, "action");
-  const args = isRecord(input.args) ? input.args : {};
-  const context = isRecord(input.context) ? input.context : {};
-  if (!extensionId || !action) {
-    throw new ApiError(400, "invalid_payload", "extensionId and action are required");
-  }
-  const registered = OPENWORK_EXPERIMENTAL_EXTENSION_ACTIONS.find((item) => item.extensionId === extensionId && item.action === action);
-  if (!registered) {
-    throw new ApiError(404, "extension_action_not_found", "OpenWork extension action not found");
-  }
-
-  if (extensionId === GOOGLE_WORKSPACE_EXTENSION_ID) {
-    const result = await callGoogleWorkspaceExtensionAction(config, action, args, context);
-    if (result) return result;
-  }
-
-  throw new ApiError(501, "extension_action_not_implemented", `${registered.title} is registered but not implemented on openwork-server yet.`, { extensionId, action, args });
 }
 
 function readStringField(value: unknown, key: string): string {
